Working on the MP votes

+++ LAB2/utils.py
@@ -140,7 +140,7 @@ def load_binary_sequence_to_matrix(file, rows=32, cols=84):
     with open(file, 'r') as file:
         sequence = file.read().replace('\n', '')  # Remove newlines if present
     # Convert the string to a list of integers
-    binary_list = [int(bit) for bit in sequence.split(',')]
+    binary_list = [float(bit) for bit in sequence.split(',')]
     
     # Convert the list to a numpy array and reshape into the desired matrix
     binary_matrix = np.array(binary_list).reshape(rows, cols)
